Summary:
Merge pg15 branch commit c1ce2c172c0b08ab6330726cb25658c0922352ba titled,
  [pg15] merge: master branch commit 'f9aad77568aac1162e1c71c1e45c34072c50d66f' into pg15
and committed 2024-06-14T14:17:59-07:00 into branch pg15-upgrade.

Clean merge with no conflicts.

Test Plan:
On MacOS 14 arm64:
  ./yb_build.sh release --sj

On AlmaLinux 8:
  ./yb_build.sh fastdebug --sj
  pg15_tests/run_tests.sh

Notes about AlmaLinux 8:
 * Upgrade tests don’t work in debug builds due to a known issue
 * org.yb.pgsql.TestPgReplicationSlot must be run with TZ=UTC (issue #22638)

Reviewers: jason

Reviewed By: jason

Subscribers: jason, ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D35902

+++ managed/src/main/java/com/yugabyte/yw/controllers/TelemetryProviderController.java
@@ -13,6 +13,7 @@ import com.yugabyte.yw.models.Audit;
 import com.yugabyte.yw.models.Customer;
 import com.yugabyte.yw.models.TelemetryProvider;
 import com.yugabyte.yw.models.common.YbaApi;
+import com.yugabyte.yw.models.helpers.CommonUtils;
 import com.yugabyte.yw.models.helpers.TelemetryProviderService;
 import com.yugabyte.yw.rbac.annotations.AuthzPath;
 import com.yugabyte.yw.rbac.annotations.PermissionAttribute;
@@ -26,12 +27,15 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.Authorization;
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
+import lombok.extern.slf4j.Slf4j;
 import play.mvc.Http;
 import play.mvc.Result;
 
 @Api(
     value = "Telemetry Provider",
     authorizations = @Authorization(AbstractPlatformController.API_KEY_AUTH))
+@Slf4j
 public class TelemetryProviderController extends AuthenticatedController {
 
   @Inject private TelemetryProviderService telemetryProviderService;
@@ -51,7 +55,7 @@ public class TelemetryProviderController extends AuthenticatedController {
     Customer.getOrBadRequest(customerUUID);
     TelemetryProvider provider =
         telemetryProviderService.getOrBadRequest(customerUUID, providerUUID);
-    return PlatformResults.withData(provider);
+    return PlatformResults.withData(CommonUtils.maskObject(provider));
   }
 
   @ApiOperation(
@@ -69,7 +73,10 @@ public class TelemetryProviderController extends AuthenticatedController {
   })
   public Result listTelemetryProviders(UUID customerUUID) {
     Customer.getOrBadRequest(customerUUID);
-    List<TelemetryProvider> providers = telemetryProviderService.list(customerUUID);
+    List<TelemetryProvider> providers =
+        telemetryProviderService.list(customerUUID).stream()
+            .map(tp -> CommonUtils.maskObject(tp))
+            .collect(Collectors.toList());
     return PlatformResults.withData(providers);
   }
 
@@ -121,7 +128,28 @@ public class TelemetryProviderController extends AuthenticatedController {
   })
   public Result deleteTelemetryProvider(
       UUID customerUUID, UUID providerUUID, Http.Request request) {
-    Customer.getOrBadRequest(customerUUID);
+    Customer customer = Customer.getOrBadRequest(customerUUID);
+
+    // Check if telemetry provider exists.
+    boolean doesTelemetryProviderExist =
+        telemetryProviderService.checkIfExists(customer.getUuid(), providerUUID);
+    if (!doesTelemetryProviderExist) {
+      String errorMessage =
+          String.format("Telemetry Provider '%s' does not exist.", providerUUID.toString());
+      log.error(errorMessage);
+      throw new PlatformServiceException(BAD_REQUEST, errorMessage);
+    }
+
+    // Check if telemetry provider is in use.
+    boolean isProviderInUse = telemetryProviderService.isProviderInUse(customer, providerUUID);
+    if (isProviderInUse) {
+      String errorMessage =
+          String.format(
+              "Cannot delete Telemetry Provider '%s', as it is in use.", providerUUID.toString());
+      log.error(errorMessage);
+      throw new PlatformServiceException(BAD_REQUEST, errorMessage);
+    }
+
     telemetryProviderService.delete(providerUUID);
     auditService()
         .createAuditEntryWithReqBody(
